@@ -2,7 +2,11 @@
     "name": "glue42-configuration-validator",
     "displayName": "Glue42 configuration validator",
     "description": "Tool which lints the application configuration files for glue",
-    "version": "0.0.5",
+    "license": "./LICENSE",
+    "version": "0.2.6",
+    "repository": {
+        "url": "https://github.com/Tick42/vscode-glue42-app-config-validator"
+    },
     "engines": {
         "vscode": "^1.29.0"
     },
@@ -61,9 +65,9 @@
                     "default": "%LOCALAPPDATA%/Tick42/GlueDesktop/config/apps",
                     "description": "Path to to the folder that Glue Desktop uses for application configurations"
                 },
-                "glue42-configuration-validator.underlineErrors":{
-                    "type":"boolean",
-                    "default":true,
+                "glue42-configuration-validator.underlineErrors": {
+                    "type": "boolean",
+                    "default": true,
                     "description": "Choose whether to underline any configuration errors in the editor"
                 }
             }
@@ -73,25 +77,29 @@
         "vscode:prepublish": "npm run compile",
         "compile": "tsc -p ./",
         "watch": "tsc -watch -p ./",
-        "postinstall": "node ./node_modules/vscode/bin/install",
+        "postinstall": "node ./node_modules/vscode/bin/install && gulp postInstall",
         "test": "npm run compile && node ./node_modules/vscode/bin/test"
     },
     "devDependencies": {
+        "@types/json-schema": "^7.0.3",
         "@types/json-stable-stringify": "^1.0.32",
+        "@types/json-to-ast": "^2.1.0",
         "@types/json5": "0.0.30",
-        "@types/lodash": "^4.14.120",
+        "@types/lodash": "^4.14.138",
         "@types/mocha": "^2.2.42",
-        "@types/node": "^8.10.25",
+        "@types/node": "^12.7.3",
+        "gulp": "^4.0.2",
         "tslint": "^5.8.0",
-        "typescript": "^3.1.4",
+        "typescript": "^3.6.2",
         "vscode": "^1.1.25"
-
     },
     "dependencies": {
-        "@types/json-to-ast": "^2.1.0",
         "ajv": "^6.4.0",
         "fs-assistant": "^0.1.2",
+        "json-fixer": "^1.3.3",
+        "json-schema-ref-parser": "^7.1.1",
         "json-to-ast": "^2.1.0",
-        "json5": "^2.1.0"
+        "json5": "^2.1.0",
+        "lodash": "^4.17.15"
     }
 }
